Encapsulate checksum handling into a module in MigrationConnector
